requirements.txt
backcall==0.1.0
decorator==4.4.0
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
ipdb==0.12
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
Markdown==3.1
mysqlclient==1.4.2.post1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
